split newsletter component from footer
